Refactoring vulnerability card code separating out style, content, and logic
