+++ frontend/src/pages/Risk/utils.ts
@@ -1,0 +1,7 @@
+export const getSeverityColor = ({ id }: { id: string }) => {
+  if (id === 'null' || id === '') return '#EFF1F5';
+  else if (id === 'Low') return '#F8DFE2';
+  else if (id === 'Medium') return '#F2938C';
+  else if (id === 'High') return '#B51D09';
+  else return '#540C03';
+};
